+++ src/components/account/DeleteAccountPageContainer.tsx
@@ -76,7 +76,10 @@ const DeleteAccountPageContainer: React.FunctionComponent<
                         : 'There was an error',
                   }
 
-                  if (error.response.status === HttpStatusCodes.UNAUTHORIZED) {
+                  if (
+                    error.response &&
+                    error.response.status === HttpStatusCodes.UNAUTHORIZED
+                  ) {
                     tokenActions.delete()
                   } else {
                     actions.setErrors(errors)
